requirements.txt
pytest~=7.4.4
selene>=2.0.0rc2
allure-pytest
pydantic~=1.10.9